Moved Resources and Bookmarks to pages folder
