+++ frontend/src/App.js
@@ -1,9 +1,9 @@
 import { Routes, Route, Navigate } from "react-router-dom";
 import "./App.css";
-import Bookmarks from "./components/Booksmarks/Bookmarks";
+import Bookmarks from "./pages/Booksmarks/Bookmarks";
 import Header from "./components/Header/Header";
 import Nav from "./components/Nav/Nav";
-import Resources from "./components/Resources/Resources";
+import Resources from "./pages/Resources/Resources";
 
 function App() {
   return (
